Bah schema.rb fix (#537)
* 4063: Setup parent/child version relationships for institutions and zip code rates

* 4063: Setup parent/child version relationships for institutions and zip code rates: Split out DB changes

* 4063: Add indexing for version_id, update AR migration version

* Remove indexing for zipcode_rates_archives and institutions_archives

* Correction to schema.rb after merge of bah-4063-db

* change version_id to bigint

+++ db/migrate/20191211111001_add_version_id_column_to_zipcode_rates_archives.rb
@@ -1,5 +1,5 @@
 class AddVersionIdColumnToZipcodeRatesArchives < ActiveRecord::Migration[5.2]
   def change
-    add_column :zipcode_rates_archives, :version_id, :integer
+    add_column :zipcode_rates_archives, :version_id, :bigint
   end
 end
